new hartree average for extrapolation
Strictly speaking, wave function extrapolation is allowed only when the
(Hartree) potential has already converged to a constant.
This is often difficult to achieve, so we usually simply took the
average value of the Hartree potential.
However, in cells with large amounts of vacuum, the average value of the
Hartree potential may be quite different from the value of the Hartree
potential where the electrons are. For this reason, we now introduce an
option to weigh the average value of the Hartree potential by the
electron density of the orbital that is being extrapolated.